q1/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
transformers>=4.36.0
torch>=2.2.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
pandas>=2.2.0
matplotlib>=3.8.0 